requirements.txt
azure-storage-blob
python-dotenv
azure-ai-formrecognizer
openai
azure.cosmos
aiohttp